All API endpoints now set the Cache-Control header to support browser caching.
Most endpoints will currently use a expiration timeout of 24h (configurable in settings.py). Three endpoints (today/ago/in) are dependent from 'today in UTC', so they will set the expiration to the end of the same day in UTC (unless the configured maximum timeout is shorter than that, in which case that is used).